Run post_init_hook after initialization is complete
The post_init_hook may be specified when instantiating JuliaProject.
It is called at the end of `ensure_init`. With this PR, the hook is called
just after the initialization is marked as complete. That is, it is not considered
part of initialization.

Before this PR, the hook was considered part of initialization, so the initialization was not flagged
as successful until after it ran. But, the hook may (and in our example
code for myjuliamod, does) call ensure_init. This resulted in a nested call
and an error message because ensure_init is called when initialization is
started but not complete.

julia_project issue #13 reports the same error message, and it looks like it might have
the same origin. However, qiskit_alt does not AFAICT call ensure_init in this nested way.
So, the present commit may not fix #13.